# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: XWiki API
  version: 0.8.0
  description: |
    面向人类与 AI Agent 的 Git-backed 文档管理系统。
    写入端点支持 Idempotency-Key 幂等。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: 登录（session cookie）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      summary: 当前用户
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/password:
    post:
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /import/bundle:
    post:
      summary: 导入 Git Bundle 创建项目（multipart file）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBundleResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /import/repo:
    post:
      summary: 从远程 Git 仓库导入创建项目
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
        - name: url
          in: query
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBundleResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users:
    get:
      summary: 用户列表（管理员）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: 创建用户（管理员）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/{id}/disable:
    post:
      summary: 禁用用户（管理员）
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          $ref: "#/components/responses/UserResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/enable:
    post:
      summary: 启用用户（管理员）
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          $ref: "#/components/responses/UserResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}:
    delete:
      summary: 删除用户（管理员，会话级联清除）
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /tokens:
    post:
      summary: 创建 Agent Token（secret 仅此一次返回）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [token, secret]
                properties:
                  token:
                    $ref: "#/components/schemas/Token"
                  secret:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      summary: Token 列表
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/Token"
  /tokens/{id}:
    delete:
      summary: 撤销 Token（幂等）
      parameters:
        - $ref: "#/components/parameters/TokenID"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /projects:
    post:
      summary: 创建项目（自动初始化 Git 仓库与 README）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          $ref: "#/components/responses/ProjectResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      summary: 项目列表（含已归档，按创建时间倒序）
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [deleted] }
          description: 仅用户会话可用；status=deleted 返回软删除项目
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
  /projects/import-folder:
    post:
      summary: 从本地文件夹导入创建项目（multipart files）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, files]
              properties:
                name:
                  type: string
                description:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          $ref: "#/components/responses/ProjectResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}:
    get:
      summary: 项目详情
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResult"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: 重命名项目并更新 README（仅用户会话）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/ProjectResult"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      summary: 软删除项目（仅用户会话，可恢复）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/archive:
    post:
      summary: 归档项目（幂等）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/unarchive:
    post:
      summary: 恢复归档项目（幂等）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/restore:
    post:
      summary: 恢复软删除项目（仅用户会话）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/purge:
    post:
      summary: 从全部 Git 历史清除指定路径（仅用户会话）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paths]
              properties:
                paths:
                  type: array
                  items: { type: string }
                message: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      summary: 永久删除已软删除项目（仅用户会话）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{id}/docs/tree:
    get:
      summary: 文档树（Git Tree 单层）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: path
          in: query
          required: false
          schema: { type: string }
          description: 目录路径，省略或空为仓库根
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [path, tree]
                properties:
                  path:
                    type: string
                  tree:
                    type: array
                    items:
                      $ref: "#/components/schemas/TreeEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/docs/home:
    get:
      summary: 项目首页（README 渲染 HTML）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/DocPage"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /projects/{id}/revision:
    get:
      summary: 当前 revision（HEAD，写入 base）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/changesets:
    post:
      summary: 原子提交 ChangeSet（可带 Idempotency-Key）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: dry_run
          in: query
          required: false
          schema: { type: boolean }
          description: 只预览不写 ref
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangesetRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangesetResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /projects/{id}/locks:
    get:
      summary: 锁状态
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          $ref: "#/components/responses/LockResult"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: 获取页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          $ref: "#/components/responses/LockResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      summary: 释放页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [released]
                properties:
                  released:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/locks/heartbeat:
    post:
      summary: 续租页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          $ref: "#/components/responses/LockResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{id}/locks/force-release:
    post:
      summary: 强制释放页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [released]
                properties:
                  released:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/shares:
    post:
      summary: 创建页面分享链接（同页幂等复用）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/search:
    get:
      summary: 项目内全文搜索
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: q
          in: query
          required: true
          schema: { type: string, maxLength: 200 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, maximum: 50, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [query, results]
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/backlinks:
    get:
      summary: 指向某文档的反向链接
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [path, backlinks]
                properties:
                  path:
                    type: string
                  backlinks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Backlink"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/export.zip:
    get:
      summary: 导出项目工作树 ZIP 快照
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: ZIP 文件
          content:
            application/zip: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/export.bundle:
    get:
      summary: 导出 Git Bundle（全仓历史）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Git bundle
          content:
            application/octet-stream: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/import:
    post:
      summary: 批量导入文件（单次原子提交）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
...
  /projects/{id}/audit:
    get:
      summary: 审计日志（session，最近 50 条）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
  /projects/{id}/commits:
    get:
      summary: 全 refs Commit 搜索与分页（倒序）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: limit
          in: query
          required: false
          schema: { type: integer, maximum: 100, default: 20 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0 }
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
          description: 匹配提交消息、作者、完整或短 SHA
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [commits, has_more]
                properties:
                  commits:
                    type: array
                    items:
                      $ref: "#/components/schemas/Commit"
                  has_more:
                    type: boolean
  /projects/{id}/commits/{sha}:
    get:
      summary: Commit 详情
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/CommitSHA"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [commit]
                properties:
                  commit:
                    $ref: "#/components/schemas/CommitDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/commits/{sha}/diff:
    get:
      summary: 机器可读 Diff（numstat/patch）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/CommitSHA"
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [numstat, patch], default: numstat }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DiffStats"
                  - $ref: "#/components/schemas/DiffPatch"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/commits/{sha}/revert:
    post:
      summary: Revert 提交（新提交，原历史保留）
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/CommitSHA"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [commit]
                properties:
                  commit:
                    $ref: "#/components/schemas/Commit"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xwiki_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      schema: { type: string }
    UserID:
      name: id
      in: path
      required: true
      schema: { type: string }
    TokenID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok:
                type: boolean
    BadRequest:
      description: 参数或校验错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证或凭据错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 权限不足（admin_required / agent_forbidden / account_disabled）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 冲突（revision_conflict / page_locked / idempotency_conflict / 名称冲突）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ProjectResult:
      description: OK
      content:
        application/json:
          schema:
            type: object
            required: [project]
            properties:
              project:
                $ref: "#/components/schemas/Project"
    UserResult:
      description: OK
      content:
        application/json:
          schema:
            type: object
            required: [user]
            properties:
              user:
                $ref: "#/components/schemas/User"
    DocPage:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DocPage"
...
  schemas:
...
    User:
      type: object
      required: [id, username, display_name, is_admin, disabled, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        is_admin:
          type: boolean
        disabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    PasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        display_name:
          type: string
        is_admin:
          type: boolean
          default: false
    Token:
      type: object
      required: [id, name, scope, project_ids, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        scope:
          type: string
          enum: [read, write]
        project_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
    TokenCreateRequest:
      type: object
      required: [name, scope, project_ids]
      properties:
        name:
          type: string
        scope:
          type: string
          enum: [read, write]
        project_ids:
          type: array
          items:
            type: string
    Project:
      type: object
      required:
        [
          id,
          name,
          description,
          repo_dir,
          archived,
          deleted,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        repo_dir:
          type: string
        archived:
          type: boolean
        deleted:
          type: boolean
        deleted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,63}$"
        description:
          type: string
    TreeEntry:
      type: object
      required: [name, type, path]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [blob, tree]
        path:
          type: string
        sha:
          type: string
...
    Revision:
      type: object
      required: [revision]
      properties:
        revision:
          type: string
...
    ImportBundleResult:
      type: object
      required: [project, commits]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        commits:
          type: integer