# internal/server/router.go
package server

import (
	"database/sql"
	"io/fs"
	"log/slog"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/cors"

	"xwiki/internal/agent"
	"xwiki/internal/auth"
	"xwiki/internal/config"
	"xwiki/internal/httpapi"
	"xwiki/internal/httpapi/handlers"
	"xwiki/internal/httpapi/middleware"
	"xwiki/internal/httpapi/response"
	"xwiki/internal/project"
	"xwiki/internal/search"
	"xwiki/internal/user"
	"xwiki/web"
)

func NewRouter(cfg *config.Config, log *slog.Logger, db *sql.DB, users *user.Store, authSvc *auth.Service, projectsSvc *project.Service, agentSvc *agent.Service, searchSvc *search.Service) http.Handler {
	r := chi.NewRouter()

	r.Use(middleware.RequestID)
	r.Use(middleware.RequestLogger(log))
	r.Use(middleware.Recoverer(log))
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{cfg.WebOrigin},
		AllowedMethods:   []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Content-Type", "Authorization", "Idempotency-Key", "X-Request-ID"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	h := handlers.NewAuthHandler(cfg, authSvc, users, log)
	ph := handlers.NewProjectHandler(cfg, projectsSvc, searchSvc, agentSvc, log)
	dh := handlers.NewDocsHandler(cfg, projectsSvc, agentSvc, log)
	ch := handlers.NewChangesetHandler(cfg, projectsSvc, agentSvc, searchSvc, log)
	th := handlers.NewTokenHandler(cfg, agentSvc, log)
	xh := handlers.NewTransferHandler(cfg, projectsSvc, agentSvc, searchSvc, log)
	ah := handlers.NewAttachmentHandler(cfg, projectsSvc, agentSvc, log)
	uh := handlers.NewUserHandler(cfg, authSvc, users, log)
	gh := handlers.NewGitHTTPHandler(projectsSvc, agentSvc, searchSvc, log)
	lh := handlers.NewLockHandler(db, agentSvc, log)
	hh := handlers.NewHistoryHandler(cfg, projectsSvc, searchSvc, agentSvc, log)
	sh := handlers.NewSearchHandler(cfg, searchSvc, projectsSvc, agentSvc, log)
	shareH := handlers.NewShareHandler(db, dh, projectsSvc, log)

	r.Get("/api/openapi.json", httpapi.OpenAPIHandler)
	r.Get("/healthz", func(w http.ResponseWriter, r *http.Request) {
		response.WriteJSON(w, http.StatusOK, map[string]any{"status": "ok"})
	})
	r.Get("/readyz", func(w http.ResponseWriter, r *http.Request) {
		if err := db.PingContext(r.Context()); err != nil {
			response.WriteError(w, r, http.StatusServiceUnavailable, "not_ready", "database unavailable")
			return
		}
		response.WriteJSON(w, http.StatusOK, map[string]any{"status": "ready"})
	})

	r.Route("/api/v1", func(r chi.Router) {
		r.Get("/meta", httpapi.MetaHandler)
		r.NotFound(func(w http.ResponseWriter, r *http.Request) {
			response.WriteError(w, r, http.StatusNotFound, "not_found", "resource not found")
		})
		r.Route("/auth", func(r chi.Router) {
			r.Post("/login", h.Login)
			r.Post("/logout", h.Logout)
			r.Post("/forgot-password", h.ForgotPassword)
			r.Post("/reset-password", h.ResetPassword)
			r.Group(func(r chi.Router) {
				r.Use(middleware.SessionAuth(authSvc))
				r.Get("/me", h.Me)
				r.Post("/password", h.Password)
			})
		})
		r.Group(func(r chi.Router) {
			r.Use(middleware.SessionAuth(authSvc))
			r.Post("/import/bundle", xh.ImportBundle)
			r.Post("/import/repo", xh.ImportRepo)
		})
		r.Route("/users", func(r chi.Router) {
			r.Group(func(r chi.Router) {
				r.Use(middleware.SessionAuth(authSvc))
				r.Use(middleware.AdminOnly)
				r.Get("/", uh.List)
				r.Post("/", uh.Create)
				r.Post("/{id}/disable", uh.Disable)
				r.Post("/{id}/enable", uh.Enable)
				r.Delete("/{id}", uh.Delete)
			})
		})
		r.Route("/tokens", func(r chi.Router) {
			r.Group(func(r chi.Router) {
				r.Use(middleware.SessionAuth(authSvc))
				r.Post("/", th.Create)
				r.Get("/", th.List)
				r.Delete("/{id}", th.Revoke)
			})
		})
		r.Route("/projects", func(r chi.Router) {
			r.Group(func(r chi.Router) {
				r.Use(middleware.AgentAuth(agentSvc))
				r.Use(middleware.SessionAuth(authSvc))
				projectAccess := middleware.AgentProjectAccess(agentSvc)
				r.With(middleware.HumanOnly).Post("/", ph.Create)
				r.With(middleware.HumanOnly).Post("/import-folder", ph.ImportFolder)
				r.Get("/", ph.List)
				r.With(projectAccess).Get("/{id}", ph.Get)
				r.With(middleware.HumanOnly).Patch("/{id}", ph.Rename)
				r.With(middleware.HumanOnly).Delete("/{id}", ph.Delete)
				r.With(middleware.HumanOnly).Post("/{id}/archive", ph.Archive)
				r.With(middleware.HumanOnly).Post("/{id}/unarchive", ph.Unarchive)
				r.With(middleware.HumanOnly).Post("/{id}/restore", ph.RestoreDeleted)
				r.With(middleware.HumanOnly).Delete("/{id}/purge", ph.PurgeDeleted)
				r.With(middleware.HumanOnly).Post("/{id}/purge", ph.Purge)
				r.With(projectAccess).Get("/{id}/docs/tree", dh.Tree)
				r.With(projectAccess).Get("/{id}/docs/home", dh.Home)
				r.With(projectAccess).Get("/{id}/docs/pages/*", dh.Page)
				r.With(projectAccess).Get("/{id}/revision", ch.Revision)
				r.With(projectAccess).Post("/{id}/changesets", ch.Apply)
				r.With(projectAccess).Get("/{id}/locks", lh.Status)
				r.With(projectAccess).Post("/{id}/locks", lh.Acquire)
				r.With(projectAccess).Delete("/{id}/locks", lh.Release)
				r.With(projectAccess).Post("/{id}/locks/heartbeat", lh.Heartbeat)
				r.With(middleware.HumanOnly).Post("/{id}/locks/force-release", lh.ForceRelease)
				r.With(middleware.HumanOnly).Post("/{id}/shares", shareH.Create)
				r.With(projectAccess).Get("/{id}/search", sh.Search)
				r.With(projectAccess).Get("/{id}/backlinks", sh.Backlinks)
				r.With(middleware.HumanOnly).Get("/{id}/export.zip", xh.ExportZip)
				r.With(middleware.HumanOnly).Get("/{id}/export.bundle", xh.ExportBundle)
				r.With(middleware.HumanOnly).Post("/{id}/import", xh.Import)
				r.With(projectAccess).Get("/{id}/attachments/*", ah.Download)
				r.With(middleware.HumanOnly).Get("/{id}/audit", th.Audit)
				r.With(projectAccess).Get("/{id}/commits", hh.Commits)
				r.With(projectAccess).Get("/{id}/commits/{sha}", hh.Commit)
				r.With(projectAccess).Get("/{id}/commits/{sha}/diff", hh.Diff)
				r.With(middleware.HumanOnly).Post("/{id}/commits/{sha}/revert", hh.Revert)
				r.With(projectAccess).Get("/{id}/files/history/*", hh.FileHistory)
			})
		})
	})

	// Public per-page share links: /share/{token} renders a standalone page.
	r.Get("/share/{token}", shareH.View)

	// Docs URL is a SPA route for browsers, but agents/crawlers/curl get a
	// server-rendered HTML page so a shared docs link "just works" without
	// JavaScript (e.g. Claude fetching the URL directly).
	spa := spaHandler()
	r.Get("/projects/{id}/docs/*", func(w http.ResponseWriter, r *http.Request) {
		if isBrowserAgent(r) {
			spa.ServeHTTP(w, r)
			return
		}
		dh.ServeView(w, r)
	})
	r.Handle("/git/{projectID}/*", gh)
	r.Handle("/*", spa)
	return r
}

// isBrowserAgent reports whether the client is a real browser (keep the SPA)
// vs an agent/crawler/CLI that should get server-rendered HTML.
func isBrowserAgent(r *http.Request) bool {
	ua := strings.ToLower(r.UserAgent())
	if ua == "" {
		return false
	}
	for _, marker := range []string{"bot", "spider", "crawler", "curl", "wget", "python", "node", "claude", "gptbot"} {
		if strings.Contains(ua, marker) {
			return false
		}
	}
	return strings.Contains(ua, "mozilla")
}

// spaHandler serves the embedded frontend with SPA fallback; unknown /api
// paths get a JSON 404 instead of the SPA index.
func spaHandler() http.Handler {
	sub, err := fs.Sub(web.Dist, "dist")
	if err != nil {
		panic(err)
	}
	files := http.FileServer(http.FS(sub))
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if strings.HasPrefix(r.URL.Path, "/api/") {
			response.WriteError(w, r, http.StatusNotFound, "not_found", "resource not found")
			return
		}
		path := strings.TrimPrefix(r.URL.Path, "/")
		if path == "" {
			path = "index.html"
		}
		if _, err := fs.Stat(sub, path); err != nil {
			r.URL.Path = "/"
		}
		files.ServeHTTP(w, r)
	})
}